FIX Location page - Update to Build-006
Tried a fix for the Location page loading properly. There may still be
infinite loops in the calls for loading the separate layers which is
causing the page to crash. It currently loads and operates without
issues but might still cause problems with slower data speeds.

Pushed build to TestFlight